+++ src/onemod/backend/jobmon_backend.py
@@ -75,6 +75,7 @@ def evaluate_with_jobmon(
     paramsets: dict[str, Any | list[Any]] | None = None,
     collect: bool | None = None,
     task_prefix: str | None = None,
+    task_attributes: dict[str, Any] = dict(),
     template_prefix: str | None = None,
     max_attempts: int = 1,
     **kwargs,
@@ -122,6 +123,10 @@ def evaluate_with_jobmon(
     task_prefix : str, optional
         Optional prefix to append to task names. Default is None, no
         prefix.
+    task_attributes : dict
+        Dictionary containing task attribute names and values. Note that
+        the task attributes will be shared across all tasks in any given
+        Workflow. Default is an empty dict, no task attributes.
     template_prefix : str, optional
         Optional prefix to append to task template name. Default is None,
         no prefix.
@@ -147,6 +152,7 @@ def evaluate_with_jobmon(
         paramsets=paramsets,
         collect=collect,
         task_prefix=task_prefix,
+        task_attributes=task_attributes,
         template_prefix=template_prefix,
         max_attempts=max_attempts,
         **kwargs,
@@ -166,6 +172,7 @@ def add_tasks_to_workflow(
     paramsets: dict[str, Any | list[Any]] | None = None,
     collect: bool | None = None,
     task_prefix: str | None = None,
+    task_attributes: dict[str, Any] = dict(),
     template_prefix: str | None = None,
     max_attempts: int = 1,
     external_upstream_tasks: list[Task] | None = None,
@@ -219,6 +226,10 @@ def add_tasks_to_workflow(
     task_prefix : str, optional
         Optional prefix to append to task names. Default is None, no
         prefix.
+    task_attributes : dict
+        Dictionary containing task attribute names and values. Note that
+        the task attributes will be shared across all tasks in any given
+        Workflow. Default is an empty dict, no task attributes.
     template_prefix : str, optional
         Optional prefix to append to task template name. Default is None,
         no prefix.
@@ -247,6 +258,7 @@ def add_tasks_to_workflow(
         paramsets=paramsets,
         collect=collect,
         task_prefix=task_prefix,
+        task_attributes=task_attributes,
         template_prefix=template_prefix,
         max_attempts=max_attempts,
         external_upstream_tasks=external_upstream_tasks,
@@ -346,6 +358,7 @@ def get_tasks(
     paramsets: dict[str, Any | list[Any]] | None,
     collect: bool | None,
     task_prefix: str | None,
+    task_attributes: dict[str, Any],
     template_prefix: str | None,
     max_attempts: int,
     external_upstream_tasks: list[Task] | None = None,
@@ -390,6 +403,8 @@ def get_tasks(
     -----------------
     task_prefix : str, optional
         Optional prefix to append to task names.
+    task_attributes : dict, optional
+        Optional dictionary containing task attribute names and values.
     template_prefix : str, optional
         Optional prefix to append to task template name.
     max_attempts : int
@@ -415,6 +430,7 @@ def get_tasks(
             stages=stages,
             external_upstream_tasks=external_upstream_tasks,
             task_prefix=task_prefix,
+            task_attributes=task_attributes,
             template_prefix=template_prefix,
             max_attempts=max_attempts,
             **kwargs,
@@ -426,6 +442,7 @@ def get_tasks(
         resources=resources,
         python=python,
         task_prefix=task_prefix,
+        task_attributes=task_attributes,
         template_prefix=template_prefix,
         max_attempts=max_attempts,
         subsets=subsets,
@@ -445,6 +462,7 @@ def get_pipeline_tasks(
     stages: list[str] | None,
     external_upstream_tasks: list[Task] | None,
     task_prefix: str | None,
+    task_attributes: dict[str, Any],
     template_prefix: str | None,
     max_attempts: int,
     **kwargs,
@@ -471,6 +489,8 @@ def get_pipeline_tasks(
         should be treated as upstream dependencies of the new tasks.
     task_prefix : str, optional
         Optional prefix to append to task names.
+    task_attributes : dict, optional
+        Optional dictionary containing task attribute names and values.
     template_prefix : str, optional
         Optional prefix to append to task template name.
     max_attempts : int
@@ -492,7 +512,13 @@ def get_pipeline_tasks(
         stage = pipeline.stages[stage_name]
         if method not in stage.skip:
             upstream_tasks = get_upstream_tasks(
-                stage, method, pipeline.stages, task_dict, stages
+                stage=stage,
+                method=method,
+                stage_dict=pipeline.stages,
+                task_dict=task_dict,
+                stages=stages,
+                task_prefix=task_prefix,
+                template_prefix=template_prefix,
             )
             task_dict[stage_name] = get_stage_tasks(
                 stage=stage,
@@ -501,6 +527,7 @@ def get_pipeline_tasks(
                 resources=resources,
                 python=python,
                 task_prefix=task_prefix,
+                task_attributes=task_attributes,
                 template_prefix=template_prefix,
                 max_attempts=max_attempts,
                 upstream_tasks=upstream_tasks,
@@ -517,6 +544,8 @@ def get_upstream_tasks(
     stage_dict: dict[str, Stage],
     task_dict: dict[str, list[Task]],
     stages: list[str] | None,
+    task_prefix: str | None,
+    template_prefix: str | None,
 ) -> list[Task]:
     """Get upstream tasks for current stage.
 
@@ -533,6 +562,10 @@ def get_upstream_tasks(
     stages : list of str or None
         Names of all pipeline stages being evaluated. If None, assume
         all stages are being evaluated.
+    task_prefix : str, optional
+        Optional prefix to filter upstream tasks with.
+    template_prefix : str, optional
+        Optional prefix to filter task templates with.
 
     Returns
     -------
@@ -557,15 +590,32 @@ def get_upstream_tasks(
             continue
 
         upstream_stage = stage_dict[upstream_name]
+
+        # Filter possible upstreams using task and template prefixes
+        possible_upstream_tasks = task_dict[upstream_name]
+        if task_prefix:
+            possible_upstream_tasks = [
+                possible_upstream_task
+                for possible_upstream_task in possible_upstream_tasks
+                if possible_upstream_task.name.startswith(task_prefix)
+            ]
+        if template_prefix:
+            possible_upstream_tasks = [
+                possible_upstream_task
+                for possible_upstream_task in possible_upstream_tasks
+                if possible_upstream_task.node.task_template_version.task_template.template_name.startswith(
+                    template_prefix
+                )
+            ]
         if method not in upstream_stage.skip:
             if (
                 upstream_stage.has_submodels
                 and method in upstream_stage.collect_after
             ):
                 # only include task corresponding to 'collect' method
-                upstream_tasks.append(task_dict[upstream_name][-1])
+                upstream_tasks.append(possible_upstream_tasks[-1])
             else:
-                upstream_tasks.extend(task_dict[upstream_name])
+                upstream_tasks.extend(possible_upstream_tasks)
 
     # if there are no upstream tasks, add external upstream tasks
     if not upstream_tasks:
@@ -581,6 +631,7 @@ def get_stage_tasks(
     resources: dict[str, Any],
     python: Path | str,
     task_prefix: str | None,
+    task_attributes: dict[str, Any],
     template_prefix: str | None,
     max_attempts: int,
     subsets: dict[str, Any | list[Any]] | None = None,
@@ -617,6 +668,8 @@ def get_stage_tasks(
         List of upstream stage tasks. Default is None.
     task_prefix : str, optional
         Optional prefix to append to task names.
+    task_attributes : dict, optional
+        Optional dictionary containing task attribute names and values.
     template_prefix : str, optional
         Optional prefix to append to task template name.
     max_attempts : int
@@ -669,6 +722,7 @@ def get_stage_tasks(
             task_template.create_task(
                 name=task_name,
                 upstream_tasks=upstream_tasks,
+                task_attributes=task_attributes,
                 max_attempts=max_attempts,
                 entrypoint=entrypoint,
                 config=config_path,
@@ -683,6 +737,7 @@ def get_stage_tasks(
             task_template.create_task(
                 name=task_name,
                 upstream_tasks=upstream_tasks,
+                task_attributes=task_attributes,
                 max_attempts=max_attempts,
                 entrypoint=entrypoint,
                 config=config_path,
@@ -701,6 +756,7 @@ def get_stage_tasks(
                 resources=resources,
                 python=python,
                 task_prefix=task_prefix,
+                task_attributes=task_attributes,
                 template_prefix=template_prefix,
                 max_attempts=max_attempts,
                 upstream_tasks=tasks,