+++ tests/e2e/test_e2e_jobmon_backend.py
@@ -24,6 +24,7 @@ KWARGS = {
     "python": None,
     "task_prefix": "me_1234",
     "template_prefix": "jobmon_e2e_testing",
+    "task_attributes": {"modelable_entity_id": 1234},
     "max_attempts": 3,
 }
 
@@ -166,9 +167,21 @@ def test_simple_pipeline_add_tasks_to_workflow_multiple_models(
         workflow=workflow,
         method="run",
         stages=["run_1", "fit_2"],
-        **(KWARGS | {"task_prefix": "me_1235"}),
+        **(
+            KWARGS
+            | {
+                "task_prefix": "me_1235",
+                "task_attributes": {"modelable_entity_id": "1235"},
+            }
+        ),
     )
     workflow.bind()
+    for task in workflow.tasks.values():
+        assert set(task.task_attributes.keys()) == {"modelable_entity_id"}
+        assert set(task.task_attributes.values()).issubset({"1234", "1235"})
+        for upstream_task in task.upstream_tasks:
+            # Check task prefixes are always identical for upstreams
+            assert task.name[:7] == upstream_task.name[:7]
     workflow.run()
 
 
@@ -212,7 +225,19 @@ def test_parallel_pipeline_add_tasks_to_workflow_multiple_models(
         workflow=workflow,
         method="run",
         stages=["run_1", "fit_2"],
-        **(KWARGS | {"task_prefix": "me_1235"}),
+        **(
+            KWARGS
+            | {
+                "task_prefix": "me_1235",
+                "task_attributes": {"modelable_entity_id": "1235"},
+            }
+        ),
     )
     workflow.bind()
+    for task in workflow.tasks.values():
+        assert set(task.task_attributes.keys()) == {"modelable_entity_id"}
+        assert set(task.task_attributes.values()).issubset({"1234", "1235"})
+        for upstream_task in task.upstream_tasks:
+            # Check task prefixes are always identical for upstreams
+            assert task.name[:7] == upstream_task.name[:7]
     workflow.run()